* alw_jbon: (21 commits)
  bulk composition
  Minor corrections to PSQL.md.
  Move performance documentation to PSQL and update to latest design.
  Hst partitioning (#213)
  Add performance insights section.
  Version 3.0.0-alpha.6
  Switch to kotlin build files, fix int2 read, improve performance testing.
  Fix basics of the plv8 tests
  Change so that 0 is undefined, which maps empty array-buffers to undefined.
  Add text_pattern_ops to geo_grid index.
  Switch to improve JBON binary format.
  Version 3.0.0-alpha.5
  fix NakshaCollection restore from jbon
  test query fix
  not-indexed query fix
  Pm update trigger (#209)
  int size in reader fix
  published new version to artifactory
  Fix sign typo in JbReader (#206)
  codec geohash fix
  ...

+++ here-naksha-lib-plv8/src/jvmMain/kotlin/JvmPlv8ResultSet.kt
@@ -56,7 +56,7 @@ open class JvmPlv8ResultSet(private var rs: ResultSet?) {
                 "point", "line", "lseg", "box", "path", "polygon", "circle", "int4range", "int8range", "numrange",
                 "tsrange", "tstzrange", "daterange" -> row[name] = rs.getString(i)
 
-                "smallint", "int2" -> row[name] = rs.getShort(i).toInt()
+                "smallint", "int2" -> row[name] = rs.getShort(i)
                 "integer", "int4", "xid4", "oid" -> row[name] = rs.getInt(i)
                 "bigint", "int8", "xid8" -> row[name] = JvmBigInt64(rs.getLong(i))
                 "real" -> row[name] = rs.getFloat(i)
